feat: pss-cln-pses-fold-service-001 - Fold composed root into owner-private internal
